requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.0.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.3.0
orjson==3.10.18
packaging==25.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.3
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1